# Added hook support