--- a/syslog_test.go
+++ b/syslog_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package zerolog
 
 import "testing"