--- a/writer_test.go
+++ b/writer_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package zerolog
 
 import (