Hi Olivier, I found zerolog recently and it looks great! Thanks for sharing it.

I wanted to do something it didn't support, then I added hooks.

This makes it possible for the user to implement adding caller information and stack traces, satisfies #9, #11, #22, and keeps the API clean and simple.

This also supports what's being asked for on #12 (example included in this PR, see the readme file).

Let me know if I overlooked anything (such as copying the hook slice over, or resetting it somewhere).
I can provide more tests, benchmarks and update the docs if what's in this PR isn't enough.

Please let me know if this makes sense.

Thanks!