@rcoelho 2017-11-27
Here are some benchmarks, before and after the changes:

Before:

```
BenchmarkLogEmpty-8             50000000                21.4 ns/op             0 B/op          0 allocs/op
BenchmarkDisabled-8             300000000                4.20 ns/op            0 B/op          0 allocs/op
BenchmarkInfo-8                 20000000                68.4 ns/op             0 B/op          0 allocs/op
BenchmarkContextFields-8        20000000                78.6 ns/op             0 B/op          0 allocs/op
BenchmarkContextAppend-8        50000000                21.6 ns/op             0 B/op          0 allocs/op
BenchmarkLogFields-8             5000000               307 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Ints-8            20000000                80.3 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Float-8           20000000                68.6 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Times-8            1000000              1252 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Durs-8             3000000               517 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface-8        5000000               273 ns/op             224 B/op          2 allocs/op
BenchmarkLogFieldType/Interfaces-8       1000000              1045 ns/op             640 B/op          4 allocs/op
BenchmarkLogFieldType/Bools-8           20000000                62.1 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Str-8             30000000                47.2 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface(Objects)-8               1000000              1084 ns/op             640 B/op          4 allocs/op
BenchmarkLogFieldType/Bool-8                            30000000                47.8 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Floats-8                           5000000               284 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Err-8                             30000000                51.4 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Errs-8                            20000000                66.3 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Time-8                            10000000               170 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface(Object)-8               20000000               113 ns/op              48 B/op          1 allocs/op
BenchmarkLogFieldType/Int-8                             30000000                50.7 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Strs-8                            20000000                81.3 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Dur-8                             20000000                73.3 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Object-8                          20000000               110 ns/op              48 B/op          1 allocs/op
BenchmarkSamplers/BasicSampler-8                        50000000                31.5 ns/op             0 B/op          0 allocs/op
BenchmarkSamplers/RandomSampler-8                       10000000               202 ns/op               0 B/op          0 allocs/op
BenchmarkSamplers/BurstSampler-8                        50000000                35.1 ns/op             0 B/op          0 allocs/op
BenchmarkSamplers/BurstSamplerNext-8                    30000000                42.1 ns/op             0 B/op          0 allocs/op
```

After adding support for hooks:

```
BenchmarkLogEmpty-8             100000000               19.7 ns/op             0 B/op          0 allocs/op
BenchmarkDisabled-8             300000000                3.53 ns/op            0 B/op          0 allocs/op
BenchmarkInfo-8                 20000000                67.4 ns/op             0 B/op          0 allocs/op
BenchmarkContextFields-8        20000000                80.6 ns/op             0 B/op          0 allocs/op
BenchmarkContextAppend-8        100000000               26.4 ns/op             0 B/op          0 allocs/op
BenchmarkLogFields-8             5000000               263 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Ints-8            20000000                85.3 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Err-8             30000000                50.7 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Errs-8            20000000                79.6 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Dur-8             20000000                64.9 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface-8        5000000               282 ns/op             224 B/op          2 allocs/op
BenchmarkLogFieldType/Interfaces-8       1000000              1038 ns/op             640 B/op          4 allocs/op
BenchmarkLogFieldType/Bool-8            30000000                46.9 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface(Objects)-8               1000000              1136 ns/op             640 B/op          4 allocs/op
BenchmarkLogFieldType/Floats-8                           5000000               294 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Float-8                           20000000                70.6 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Strs-8                            20000000                84.7 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Time-8                            10000000               165 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Times-8                            1000000              1372 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Durs-8                             3000000               484 ns/op               0 B/op          0 allocs/op
BenchmarkLogFieldType/Interface(Object)-8               10000000               125 ns/op              48 B/op          1 allocs/op
BenchmarkLogFieldType/Int-8                             30000000                44.7 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Str-8                             30000000                52.5 ns/op             0 B/op          0 allocs/op
BenchmarkLogFieldType/Object-8                          20000000               114 ns/op              48 B/op          1 allocs/op
BenchmarkLogFieldType/Bools-8                           20000000                63.5 ns/op             0 B/op          0 allocs/op
BenchmarkSamplers/BasicSampler-8                        50000000                32.0 ns/op             0 B/op          0 allocs/op
BenchmarkSamplers/RandomSampler-8                       10000000               175 ns/op               0 B/op          0 allocs/op
BenchmarkSamplers/BurstSampler-8                        30000000                40.2 ns/op             0 B/op          0 allocs/op
BenchmarkSamplers/BurstSamplerNext-8                    30000000                41.8 ns/op             0 B/op          0 allocs/op
```

I extracted the new hook benchmarks:

```
BenchmarkHooks/Nop/Single-8                             30000000                67.3 ns/op            40 B/op          2 allocs/op
BenchmarkHooks/Nop/Multi-8                              20000000                83.5 ns/op            80 B/op          3 allocs/op
BenchmarkHooks/Simple-8                                 20000000                92.9 ns/op            40 B/op          2 allocs/op
```

There's a lot of variation between benchmark runs, though.