@rs 2017-11-27
Can we close this in favor of #25?